Cleanup after moving connections to pipeline view
