+++ src/web/pipelines/PipelineConnectionsActions.js
