+++ src/web/pipelines/PipelineConnectionsStore.js
