+++ src/web/rules/Rule.jsx
@@ -45,10 +45,6 @@ const Rule = React.createClass({
           </span>
 
           <span>
-            <LinkContainer to={Routes.pluginRoute('SYSTEM_CONNECTIONS')}>
-              <Button bsStyle="info">Manage connections</Button>
-            </LinkContainer>
-            &nbsp;
             <LinkContainer to={Routes.pluginRoute('SYSTEM_PIPELINES')}>
               <Button bsStyle="info">Manage pipelines</Button>
             </LinkContainer>
